refactor: Phase 1 SOLID — split Python package and extract pattern cache
Split the monolithic formatparse/__init__.py into focused submodules
(api, validation, bidirectional, custom, tz, compat) with a thin
re-export hub preserving all public imports. Add post_parse_validate
to deduplicate parse/ValidatedParser validation paths.

Extract LRU pattern caching from lib.rs into pattern_cache.rs and
remove unused duplicate match.rs.
